third_party/blink/web_tests/platform/generic/external/wpt/credential-management/passwordcredential-framed-get.sub.https-expected.txt

This is a testharness.js-based test.
PASS Same-origin IFrame does not throw.
PASS Cross-origin IFrame throws 'NotAllowedError'.
FAIL Same-origin IFrame in same-origin IFrame does not throw. assert_equals: expected (object) null but got (string) "NotSupportedError"
PASS Same-origin IFrame in same-origin IFrame throws 'NotAllowedError'.
PASS Cross-origin IFrame in same-origin IFrame throws 'NotAllowedError'.
PASS Cross-origin IFrame in same-cross-origin throws 'NotAllowedError'.
Harness: the test ran to completion.
